fabric-loot-api-v3
