+++ deprecated/fabric-loot-api-v2/README.md
