+++ deprecated/fabric-loot-api-v2/src/main/java/net/fabricmc/fabric/api/loot/v2/package-info.java
