+++ deprecated/fabric-loot-api-v2/src/testmod/resources/data/minecraft/predicate/match_tool_shears.json
@@ -1,0 +1,8 @@
+{
+  "condition": "minecraft:match_tool",
+  "predicate": {
+    "items": [
+      "minecraft:shears"
+    ]
+  }
+}
