+++ deprecated/fabric-loot-api-v2/src/testmod/resources/fabric.mod.json
