--- fabric-loot-api-v2/src/main/resources/fabric.mod.json
@@ -1,34 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "fabric-loot-api-v2",
-  "name": "Fabric Loot API (v2)",
-  "version": "${version}",
-  "environment": "*",
-  "license": "Apache-2.0",
-  "icon": "assets/fabric-loot-api-v2/icon.png",
-  "contact": {
-    "homepage": "https://fabricmc.net",
-    "irc": "irc://irc.esper.net:6667/fabric",
-    "issues": "https://github.com/FabricMC/fabric/issues",
-    "sources": "https://github.com/FabricMC/fabric"
-  },
-  "authors": [
-    "FabricMC"
-  ],
-  "depends": {
-    "fabricloader": ">=0.15.11",
-    "fabric-api-base": "*",
-    "fabric-resource-loader-v0": "*"
-  },
-  "description": "Hooks for manipulating loot tables.",
-  "mixins": [
-    "fabric-loot-api-v2.mixins.json"
-  ],
-  "custom": {
-    "fabric-api:module-lifecycle": "stable",
-    "loom:injected_interfaces": {
-      "net/minecraft/class_52\u0024class_53": ["net/fabricmc/fabric/api/loot/v2/FabricLootTableBuilder"],
-      "net/minecraft/class_55\u0024class_56": ["net/fabricmc/fabric/api/loot/v2/FabricLootPoolBuilder"]
-    }
-  }
-}
