+++ fabric-loot-api-v3/build.gradle
