+++ fabric-loot-api-v3/src/main/java/net/fabricmc/fabric/mixin/loot/JsonDataLoaderMixin.java
