+++ fabric-loot-api-v3/src/main/resources/assets/fabric-loot-api-v3/icon.png
