+++ fabric-loot-api-v3/src/main/resources/fabric-loot-api-v3.mixins.json
