+++ fabric-loot-api-v3/src/testmod/resources/data/minecraft/loot_table/blocks/red_wool.json
