+++ fabric-loot-api-v3/src/testmod/resources/data/minecraft/predicate/match_tool_shears.json
