+++ fabric-loot-api-v3/src/testmod/resources/fabric.mod.json
@@ -1,0 +1,16 @@
+{
+  "schemaVersion": 1,
+  "id": "fabric-loot-api-v3-testmod",
+  "name": "Fabric Loot Table API (v3) Test Mod",
+  "version": "1.0.0",
+  "environment": "*",
+  "license": "Apache-2.0",
+  "depends": {
+    "fabric-loot-api-v3": "*"
+  },
+  "entrypoints": {
+    "main": [
+      "net.fabricmc.fabric.test.loot.LootTest"
+    ]
+  }
+}
