+++ gradle.properties
@@ -36,6 +36,7 @@ fabric-key-binding-api-v1-version=1.0.47
 fabric-keybindings-v0-version=0.2.45
 fabric-lifecycle-events-v1-version=2.3.11
 fabric-loot-api-v2-version=3.0.11
+fabric-loot-api-v3-version=1.0.0
 fabric-message-api-v1-version=6.0.13
 fabric-model-loading-api-v1-version=2.0.0
 fabric-networking-api-v1-version=4.2.0
