+++ settings.gradle
@@ -37,7 +37,7 @@ include 'fabric-item-api-v1'
 include 'fabric-item-group-api-v1'
 include 'fabric-key-binding-api-v1'
 include 'fabric-lifecycle-events-v1'
-include 'fabric-loot-api-v2'
+include 'fabric-loot-api-v3'
 include 'fabric-message-api-v1'
 include 'fabric-model-loading-api-v1'
 include 'fabric-networking-api-v1'
@@ -62,6 +62,7 @@ include 'deprecated:fabric-command-api-v1'
 include 'deprecated:fabric-commands-v0'
 include 'deprecated:fabric-convention-tags-v1'
 include 'deprecated:fabric-keybindings-v0'
+include 'deprecated:fabric-loot-api-v2'
 include 'deprecated:fabric-renderer-registries-v1'
 include 'deprecated:fabric-rendering-data-attachment-v1'
 include 'deprecated:fabric-rendering-v0'
